backend/pyproject.toml
[project]
name = "ai-qa-system"
version = "3.0.0"
description = "Enterprise AI-powered QA system with LangGraph orchestration, RAG retrieval, and multi-model collaboration"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "0xxuebao"}]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "uvloop>=0.21.0",

    # LangGraph + LangChain
    "langgraph>=0.4.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",

    # LLM Providers (unified via LiteLLM)
    "litellm>=1.68.0,<=1.77.5",  # 2025 stable range, pre-supply-chain-attack

    # RAG
    "lightrag>=0.1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",

    # Redis
    "redis[hiredis]>=5.0.0",

    # Validation & Config
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # HTTP Client
    "httpx>=0.27.0",

    # SSE
    "sse-starlette>=2.0.0",

    # Monitoring
    "structlog>=24.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",

    # Utils
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "mypy>=1.10.0",
    "ruff>=0.6.0",
    "coverage>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]
